improve semantic test
