+++ components/divider/__tests__/semantic.test.tsx
@@ -6,7 +6,7 @@ import type { DividerProps } from '..';
 import mountTest from '../../../tests/shared/mountTest';
 import { render } from '../../../tests/utils';
 
-describe('Divider', () => {
+describe('Divider.Semantic', () => {
   mountTest(Divider);
 
   it('not show children when vertical', () => {
